[Feat/DetaiPage] 판매자 데이터 불러오는 컴포넌트 분리
